+++ cmake/CompilerWarnings.cmake
@@ -87,7 +87,7 @@ function(set_project_warnings project_name)
     message(AUTHOR_WARNING "No compiler warnings set for '${CMAKE_CXX_COMPILER_ID}' compiler.")
   endif()
 
-  target_compile_options(${project_name} PUBLIC ${PROJECT_WARNINGS})
+  add_compile_options(${PROJECT_WARNINGS})
 
   if(NOT TARGET ${project_name})
     message(AUTHOR_WARNING "${project_name} is not a target, thus no compiler warning were added.")